1.0.85

+++ src/config.h
@@ -70,11 +70,11 @@ extern Uint   conf_dsize	P((char*));
 extern Uint   conf_dconv	P((char*, char*, char*, Uint));
 extern void   conf_dread	P((int, char*, char*, Uint));
 
-extern void   conf_init		P((char*, char*));
-extern char  *conf_base_dir	P((void));
-extern char  *conf_driver	P((void));
-extern bool   conf_typechecking	P((void));
-extern int    conf_array_size	P((void));
+extern void		conf_init	P((char*, char*));
+extern char	       *conf_base_dir	P((void));
+extern char	       *conf_driver	P((void));
+extern bool		conf_typechecking P((void));
+extern unsigned short	conf_array_size	P((void));
 
 typedef struct _array_ array;
 typedef struct _object_ object;
